# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NVRS Menu API",
    "version": "1.0.0",
    "description": "API documentation for Node Virtual Restaurant Solutions",
    "contact": {
      "name": "Alexander Gomez"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "MenuItem": {
        "type": "object",
        "required": [
          "item_name",
          "price",
          "item_type"
        ],
        "properties": {
          "item_id": {
            "type": "integer",
            "description": "The auto-generated id of the menu item"
          },
          "item_name": {
            "type": "string",
            "description": "The name of the menu item"
          },
          "item_desc": {
            "type": "string",
            "description": "Description of the menu item"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "description": "Price of the menu item"
          },
          "item_type": {
            "type": "string",
            "description": "Type/category of the menu item"
          }
        }
      }
    }
  },
  "paths": {
    "/api/images/{imageId}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get menu item image by image ID",
        "description": "Retrieves an image from the database by its unique image ID",
        "parameters": [
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The unique identifier for the image"
          }
        ],
        "responses": {
          "200": {
            "description": "Image returned successfully",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Image not found in the database"
          },
          "500": {
            "description": "Server error occurred while retrieving the image"
          }
        }
      }
    },
    "/api/images/menu-item/{menuItemId}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get image by menu item ID",
        "description": "Retrieves the most recent image associated with a specific menu item",
        "parameters": [
          {
            "in": "path",
            "name": "menuItemId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the menu item whose image is to be retrieved"
          }
        ],
        "responses": {
          "200": {
            "description": "Image returned successfully",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "No image found for this menu item"
          },
          "500": {
            "description": "Server error occurred while retrieving the image"
          }
        }
      },
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Upload an image for a menu item",
        "description": "Uploads a new image file and associates it with a specific menu item in the database",
        "parameters": [
          {
            "in": "path",
            "name": "menuItemId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the menu item to associate with the uploaded image"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "The image file to upload (JPEG, PNG, GIF, etc.)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Image uploaded successfully"
                    },
                    "imageId": {
                      "type": "integer",
                      "example": 42
                    },
                    "menuItemId": {
                      "type": "integer",
                      "example": 15
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid file or menu item ID"
          },
          "404": {
            "description": "Menu item not found"
          },
          "500": {
            "description": "Server error occurred during upload"
          }
        }
      }
    },
    "/api/images/upload/{menuItemId}": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Upload an image for a menu item (alternative route)",
        "description": "Alternative endpoint for uploading an image file for a menu item. Functionally identical to '/api/images/menu-item/{menuItemId}'",
        "parameters": [
          {
            "in": "path",
            "name": "menuItemId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the menu item to associate with the uploaded image"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "The image file to upload (JPEG, PNG, GIF, etc.)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Image uploaded successfully"
                    },
                    "imageId": {
                      "type": "integer",
                      "example": 42
                    },
                    "menuItemId": {
                      "type": "integer",
                      "example": 15
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid file or menu item ID"
          },
          "404": {
            "description": "Menu item not found"
          },
          "500": {
            "description": "Server error occurred during upload"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "Menu Items"
        ],
        "summary": "Get all menu items",
        "description": "Retrieves all menu items from the database",
        "responses": {
          "200": {
            "description": "A list of menu items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItem"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/createMenuItem": {
      "post": {
        "tags": [
          "Menu Items"
        ],
        "summary": "Create a new menu item",
        "description": "Adds a new menu item to the database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Menu item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/menu/{id}": {
      "put": {
        "tags": [
          "Menu Items"
        ],
        "summary": "Update a menu item",
        "description": "Updates an existing menu item in the database",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Menu item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "item_name": {
                    "type": "string",
                    "description": "The name of the menu item"
                  },
                  "item_desc": {
                    "type": "string",
                    "description": "Description of the menu item"
                  },
                  "price": {
                    "type": "number",
                    "format": "decimal",
                    "description": "Price of the menu item"
                  },
                  "item_type": {
                    "type": "string",
                    "description": "Type/category of the menu item"
                  }
                },
                "example": {
                  "item_name": "Updated Burger",
                  "item_desc": "A delicious updated burger with special sauce",
                  "price": 12.99,
                  "item_type": "Main Course"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Menu item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Menu item updated successfully"
                    },
                    "itemId": {
                      "type": "integer",
                      "example": 1
                    },
                    "affectedRows": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Menu item not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Menu Items"
        ],
        "summary": "Delete a menu item",
        "description": "Deletes a menu item from the database",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Menu item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Menu item deleted successfully"
          },
          "404": {
            "description": "Menu item not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "tags": []
}